+++ AGENTS.md
@@ -13,6 +13,7 @@ This repository contains a coding assignment for building an integration layer b
 - Include a README that explains architecture, how to run and test locally, and thoughts on error handling, extensibility, security, and a short personal reflection.
 
 ## Coding guidelines
+- All projects must target **.NET 9.0**. Do not change the framework version.
 - Keep code pragmatic and minimal; favour readability similar to Go over heavy patterns.
 - Avoid reflection where possible.
 - Use comments to clarify code and tests when helpful, but keep them concise.
