+++ TpInsuranceAPI/TpInsuranceAPI.csproj
@@ -7,9 +7,10 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.8"/>
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0"/>
-        <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="9.0.0"/>
+        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="9.0.8" />
+        <PackageReference Include="Swashbuckle.AspNetCore" Version="9.0.3" />
+        <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="9.0.8" />
+        <PackageReference Include="MinApiLib.Validation" Version="9.0.0" />
     </ItemGroup>
 
 </Project>
