--- TpInsuranceAPI/TpInsuranceAPI.http
@@ -1,6 +1,0 @@
-@TpInsuranceAPI_HostAddress = http://localhost:5046
-
-GET {{TpInsuranceAPI_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
