+++ TpVehicleAPI/Contracts/VehicleResponse.cs
@@ -1,0 +1,3 @@
+namespace TpVehicleAPI;
+
+public sealed record VehicleResponse(string RegistrationNumber, string Manufacturer, string Model, int ModelYear);
