--- TpVehicleAPI/TpVehicleAPI.http
@@ -1,6 +1,0 @@
-@TpVehicleAPI_HostAddress = http://localhost:5001
-
-GET {{TpVehicleAPI_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
